+++ CHANGELOG.md
@@ -14,15 +14,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Circle2D: split_by_line
 * BoundingRectangle: bounds, plot, area, center, b_rectangle_intersection, is_inside_b_rectangle, point_belongs, intersection_area, distance_to_b_rectangle, distance_to_point
 * CylindricalSurface3D: linesegment_intersections
+* CylindricalSurface3D: line_intersections
+* CylindricalSurface3D: plane_intersection
+* Line3D: to_2d
+* Line2D: point_distance
+* 
+
 
 ### Fixed
 
 * BsplineCurve: abscissa (use different start point between 0 and length)
 * Arc3D: plot
-
+* CylindricalSurface3D: point3d_to_2d, if y < 0, add 2*pi to theta
 
 ### Performance improvements
-
 
 
 
@@ -37,6 +42,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * BsplineCurve: abscissa
 * Circle2D: split_by_line
 * BoundingRectangle: area, center, intersection, is_inside, point_belongs, intersection_area, distance_to_point, distance_to_b_rectangle
+
 
 
 ## v0.6.0 [Unrealeased]